Completed GRid tests and updated GRid to reflect first test results.
There were a number of errors in the GRid class, largely stemming from problems with the constructor and the interactions between static and non-static methods.  These have been adjusted by reverting to the base __constructor and parseCode methods inherited from Mod3736 and instead calling checkFormat at the start of methods which need it rather than on the constructor.  This allows for the class to function whether on external codes or on the code used as the object's attribute.